# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "TRH Backend API",
        "title": "TRH Backend",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:${PORT}",
    "basePath": "/api/v1",
    "paths": {
...
        "/stacks/thanos/{id}": {
            "get": {
                "description": "Get Stack By ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Thanos Stack"
                ],
                "summary": "Get Stack By ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Thanos Stack ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entities.Response"
                        }
                    }
                }
            },
            "put": {
                "description": "Update Network",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Thanos Stack"
                ],
                "summary": "Update Network",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Thanos Stack ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update Network Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.UpdateNetworkRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entities.Response"
                        }
                    }
                }
            },
            "delete": {
                "description": "Terminate Thanos Stack",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Thanos Stack"
                ],
                "summary": "Terminate Thanos Stack",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Thanos Stack ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entities.Response"
                        }
                    }
                }
            }
        },
...
        "/stacks/thanos/{id}/integrations": {
            "get": {
                "description": "Get Integrations",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Thanos Stack"
                ],
                "summary": "Get Integrations",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Thanos Stack ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entities.Response"
                        }
                    }
                }
            }
        },
        "/stacks/thanos/{id}/integrations/block-explorer": {
            "post": {
                "description": "Install Block Explorer",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Thanos Stack"
                ],
                "summary": "Install Block Explorer",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Thanos Stack ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Install Block Explorer Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.InstallBlockExplorerRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entities.Response"
                        }
                    }
                }
            },
            "delete": {
                "description": "Uninstall Block Explorer",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Thanos Stack"
                ],
                "summary": "Uninstall Block Explorer",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Thanos Stack ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entities.Response"
                        }
                    }
                }
            }
        },
        "/stacks/thanos/{id}/integrations/bridge": {
            "post": {
                "description": "Install Bridge",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Thanos Stack"
                ],
                "summary": "Install Bridge",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Thanos Stack ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entities.Response"
                        }
                    }
                }
            },
            "delete": {
                "description": "Uninstall Bridge",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Thanos Stack"
                ],
                "summary": "Uninstall Bridge",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Thanos Stack ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entities.Response"
                        }
                    }
                }
            }
        },
        "/stacks/thanos/{id}/integrations/monitoring": {
            "post": {
                "description": "Install Monitoring",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Thanos Stack"
                ],
                "summary": "Install Monitoring",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Thanos Stack ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Install Monitoring Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.InstallMonitoringRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entities.Response"
                        }
                    }
                }
            },
            "delete": {
                "description": "Uninstall Monitoring",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Thanos Stack"
                ],
                "summary": "Uninstall Monitoring",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Thanos Stack ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entities.Response"
                        }
                    }
                }
            }
        },
        "/stacks/thanos/{id}/integrations/{integrationId}": {
            "get": {
                "description": "Get Integration By ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Thanos Stack"
                ],
                "summary": "Get Integration By ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Thanos Stack ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Integration ID",
                        "name": "integrationId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entities.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "dtos.AlertManagerConfig": {
            "type": "object",
            "properties": {
                "email": {
                    "$ref": "#/definitions/dtos.EmailConfig"
                },
                "telegram": {
                    "$ref": "#/definitions/dtos.TelegramConfig"
                }
            }
        },
...
        "dtos.EmailConfig": {
            "type": "object",
            "properties": {
                "criticalReceivers": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "defaultReceivers": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "enabled": {
                    "type": "boolean"
                },
                "smtpAuthPassword": {
                    "type": "string"
                },
                "smtpAuthUsername": {
                    "type": "string"
                },
                "smtpFrom": {
                    "type": "string"
                },
                "smtpSmarthost": {
                    "type": "string"
                }
            }
        },
        "dtos.InstallBlockExplorerRequest": {
            "type": "object",
            "required": [
                "coinmarketcapKey",
                "databasePassword",
                "databaseUsername",
                "walletConnectId"
            ],
            "properties": {
                "coinmarketcapKey": {
                    "type": "string"
                },
                "databasePassword": {
                    "type": "string"
                },
                "databaseUsername": {
                    "type": "string"
                },
                "walletConnectId": {
                    "type": "string"
                }
            }
        },
        "dtos.InstallMonitoringRequest": {
            "type": "object",
            "required": [
                "grafanaPassword"
            ],
            "properties": {
                "alertManager": {
                    "$ref": "#/definitions/dtos.AlertManagerConfig"
                },
                "grafanaPassword": {
                    "type": "string"
                }
            }
        },
...
        "dtos.TelegramConfig": {
            "type": "object",
            "properties": {
                "apiToken": {
                    "type": "string"
                },
                "criticalReceivers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dtos.TelegramReceiver"
                    }
                },
                "enabled": {
                    "type": "boolean"
                }
            }
        },
        "dtos.TelegramReceiver": {
            "type": "object",
            "properties": {
                "chatId": {
                    "type": "string"
                }
            }
        },
        "dtos.UpdateNetworkRequest": {
            "type": "object",
            "properties": {
                "l1BeaconUrl": {
                    "type": "string"
                },
                "l1RpcUrl": {
                    "type": "string"
                }
            }
        },
...
        "entities.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}